CalibratedServo as a standalone class
