+++ library.properties
@@ -7,4 +7,4 @@ paragraph=Instrument panel driver for a certain horse
 category=Other
 url=https://github.com/ianfixes/ManeDisplay
 architectures=*
-includes=DashMessage.h,MasterProperties.h,SlaveProperties.h,DashState.h
+includes=CalibratedServo.h,DashMessage.h,DashState.h,MasterProperties.h,SlaveProperties.h
